# swagger.yaml
swagger: "2.0"
info:
  version: "v4"
  title: "Projects API"
# during production,should point to your server machine
host: localhost:3000
basePath: "/v4"
...
schemes:
- "http"
produces:
- application/json
consumes:
- application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /projects:
    get:
      tags:
        - project
      operationId: findProjects
      security:
        - Bearer: []
      description: Retrieve projects that match the filter
      responses:
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: A list of projects
          schema:
            $ref: "#/definitions/ProjectListResponse"
      parameters:
        - $ref: "#/parameters/offsetParam"
        - $ref: "#/parameters/limitParam"
        - name: filter
          required: true
          type: string
          in: query
          description: |
            Url encoded list of Supported filters
             - id
             - status
             - type
             - memberOnly
             - keyword
        - name: sort
          required: false
          description: |
            sort projects by status, name, type, createdAt, updatedAt. Default is createdAt asc
          in: query
          type: string
    post:
      operationId: addProject
      security:
        - Bearer: []
      description: Create a project
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/NewProjectBodyParam'
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '201':
          description: Returns the newly created project
          schema:
            $ref: "#/definitions/ProjectResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
...
  /projectTemplates:
    get:
      tags:
        - projectTemplate
      operationId: findProjectTemplates
      security:
        - Bearer: []
      description: Retrieve all project templates. All user roles can access this endpoint.
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: A list of project templates
          schema:
            $ref: "#/definitions/ProjectTemplateListResponse"
    post:
      tags:
        - projectTemplate
      operationId: addProjectTemplate
      security:
        - Bearer: []
      description: Create a project template
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ProjectTemplateBodyParam'
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '201':
          description: Returns the newly created project template
          schema:
            $ref: "#/definitions/ProjectTemplateResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
...
  /productTemplates:
    get:
      tags:
        - productTemplate
      operationId: findProductTemplates
      security:
        - Bearer: []
      description: Retrieve all product templates. All user roles can access this endpoint.
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: A list of product templates
          schema:
            $ref: "#/definitions/ProductTemplateListResponse"
    post:
      tags:
        - productTemplate
      operationId: addProductTemplate
      security:
        - Bearer: []
      description: Create a product template
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ProductTemplateBodyParam'
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '201':
          description: Returns the newly created product template
          schema:
            $ref: "#/definitions/ProductTemplateResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
...
  /projectTypes:
    get:
      tags:
        - projectType
      operationId: findProjectTypes
      security:
        - Bearer: []
      description: Retrieve all project types. All user roles can access this endpoint.
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: A list of project types
          schema:
            $ref: "#/definitions/ProjectTypeListResponse"
    post:
      tags:
        - projectType
      operationId: addProjectType
      security:
        - Bearer: []
      description: Create a project type. Only admin or connect admin can access this endpoint.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ProjectTypeCreateBodyParam'
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '201':
          description: Returns the newly created project type
          schema:
            $ref: "#/definitions/ProjectTypeResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
  /projectTypes/{key}:
    get:
      tags:
        - projectType
      description: Retrieve project type by id. All user roles can access this endpoint.
      security:
        - Bearer: []
      responses:
        '404':
          description: Not found
          schema:
            $ref: "#/definitions/ErrorModel"
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: a project type
          schema:
            $ref: "#/definitions/ProjectTypeResponse"
      parameters:
        - $ref: "#/parameters/keyParam"
      operationId: getProjectType

    patch:
      tags:
        - projectType
      operationId: updateProjectType
      security:
        - Bearer: []
      description: Update a project type. Only admin or connect admin can access this endpoint.
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '404':
          description: Not found
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: Successfully updated project type.
          schema:
            $ref: "#/definitions/ProjectTypeResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
        default:
          description: error payload
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: "#/parameters/keyParam"
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/ProjectTypeBodyParam"

    delete:
      tags:
        - projectType
      description: Remove an existing project type. Only admin or connect admin can access this endpoint.
      security:
        - Bearer: []
      parameters:
        - $ref: "#/parameters/keyParam"
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '404':
          description: If project is not found
          schema:
            $ref: "#/definitions/ErrorModel"
        '204':
          description: Project type successfully removed
  /productCategories:
    get:
      tags:
        - productCategory
      operationId: findProductCategories
      security:
        - Bearer: []
      description: Retrieve all product categories. All user roles can access this endpoint.
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: A list of product categories
          schema:
            $ref: "#/definitions/ProductCategoryListResponse"
    post:
      tags:
        - productCategory
      operationId: addProductCategory
      security:
        - Bearer: []
      description: Create a product category. Only admin or connect admin can access this endpoint.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ProductCategoryCreateBodyParam'
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '201':
          description: Returns the newly created product category
          schema:
            $ref: "#/definitions/ProductCategoryResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
  /productCategories/{key}:
    get:
      tags:
        - productCategory
      description: Retrieve product category by id. All user roles can access this endpoint.
      security:
        - Bearer: []
      responses:
        '404':
          description: Not found
          schema:
            $ref: "#/definitions/ErrorModel"
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: a product category
          schema:
            $ref: "#/definitions/ProductCategoryResponse"
      parameters:
        - $ref: "#/parameters/productCategoryKeyParam"
      operationId: getProductCategory

    patch:
      tags:
        - productCategory
      operationId: updateProductCategory
      security:
        - Bearer: []
      description: Update a product category. Only admin or connect admin can access this endpoint.
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '404':
          description: Not found
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: Successfully updated product category.
          schema:
            $ref: "#/definitions/ProductCategoryResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
        default:
          description: error payload
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: "#/parameters/productCategoryKeyParam"
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/ProductCategoryBodyParam"

    delete:
      tags:
        - productCategory
      description: Remove an existing project type. Only admin or connect admin can access this endpoint.
      security:
        - Bearer: []
      parameters:
        - $ref: "#/parameters/productCategoryKeyParam"
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '404':
          description: If project is not found
          schema:
            $ref: "#/definitions/ErrorModel"
        '204':
          description: Product category successfully removed
  /timelines:
    get:
      tags:
        - timeline
      operationId: findTimelines
      security:
        - Bearer: []
      description: Retrieve timelines which its projects are accessible by the user.
      parameters:
        - name: filter
          required: false
          type: string
          in: query
          description: |
            Url encoded list of supported filters
             - reference
             - referenceId
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: A list of timelines
          schema:
            $ref: "#/definitions/TimelineListResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
    post:
      tags:
        - timeline
      operationId: addTimeline
      security:
        - Bearer: []
      description: Create a timeline. All users who can edit the project can access this endpoint.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/TimelineBodyParam'
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '201':
          description: Returns the newly created timeline
          schema:
            $ref: "#/definitions/TimelineResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
...
parameters:
...
  keyParam:
    name: key
    in: path
    description: project type key
    required: true
    type: string
  productCategoryKeyParam:
    name: key
    in: path
    description: product category key
    required: true
    type: string
...
  offsetParam:
    name: offset
    description: "number of items to skip. Defaults to 0"
    in: query
    required: false
    type: integer
    format: int32
  limitParam:
    name: limit
    description: "max records to return. Defaults to 20"
    in: query
    required: false
    type: integer
    format: int32
definitions:
  ResponseMetadata:
    title: Metadata object for a response
    type: object
    properties:
      totalCount:
        type: integer
        format: int64
        description: Total count of the objects
  ErrorModel:
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            description: http status code
            type: integer
            format: int32
          debug:
            type: object
          content:
            type: object
...
  NewProject:
    type: object
    required:
       - name
       - description
       - type
    properties:
      name:
        type: string
        description: project name (required)
      description:
        type: string
        description: Project description
      billingAccountId:
        type: number
        format: long
        description: the customer billing account id
      estimatedPrice:
        type: number
        format: float
        description: The estimated price of the project
      terms:
        type: array
        items:
          type: number
          format: integer
      external:
        type: object
        description: READ-ONLY, OPTIONAL. Refernce to external task/issue.
        properties:
          id:
            type: string
            description: Identifier for external reference
          type:
            type: string
            description: external source type
            enum: [ "github", "jira", "asana", "other"]
          data:
            type: string
            description: "300 Char length text blob for customer provided data"
      type:
        type: string
        description: project type
      bookmarks:
        type: array
        items:
          $ref: "#/definitions/ProjectBookMark"
      challengeEligibility:
        description: List of eligibility criteria (one entry per role)
        type: array
        items:
          $ref: "#/definitions/ChallengeEligibility"
      details:
        $ref: "#/definitions/ProjectDetails"
      utm:
        description: READ-ONLY. Used for tracking
        type: object
        properties:
          campaign:
            type: string
          medium:
            type: string
          source:
            type: string
      templateId:
        description: the project template identifier
        type: number
        format: long
  NewProjectBodyParam:
    type: object
    properties:
      param:
        $ref: "#/definitions/NewProject"
...
  Project:
    type: object
    properties:
      id:
        description: unique identifier
        type: integer
        format: int64
      directProjectId:
        description: unique identifier in direct
        type: integer
        format: int64
      billingAccountId:
        type: integer
        format: int64
        description: The customer billing account id
      utm:
        description: READ-ONLY. Used for tracking
        type: object
        properties:
          campaign:
            type: string
          medium:
            type: string
          source:
            type: string
      estimatedPrice:
        type: number
        format: float
        description: The estimated price of the project
      actualPrice:
        type: number
        format: float
        description: The actual price of the project
      terms:
        type: array
        items:
          type: number
          format: integer
      name:
        type: string
        description: project name
      description:
        type: string
        description: Project description

      external:
        type: object
        description: READ-ONLY, OPTIONAL. Refernce to external task/issue.
        properties:
          id:
            type: string
            description: Identifier for external reference
          type:
            type: string
            description: external source type
            enum: [ "github", "jira", "asana", "other"]
          data:
            type: string
            description: "300 Char length text blob for customer provided data"
      type:
        type: string
        description: project type
      status:
        type: string
        description: current state of the task
        enum: ["draft", "in_review", "reviewed", "active", "paused", "cancelled", "completed"]
      cancelReason:
        type: string
        description: If a project is cancelled, define the reason of cancellation
      challengeEligibility:
        description: List of eligibility criteria (one entry per role)
        type: array
        items:
          $ref: "#/definitions/ChallengeEligibility"
      bookmarks:
        type: array
        items:
          $ref: "#/definitions/ProjectBookMark"
      members:
        description: |
          READ-ONLY. List of project members.
          Use project member api to add/remove members
        type: array
        items:
          $ref: "#/definitions/ProjectMember"
      attachments:
        description: |
          READ-ONLY. List of project attachmens.
          Use project attachment api to add/remove attachments
        type: array
        items:
          $ref: "#/definitions/ProjectAttachment"
      details:
        $ref: "#/definitions/ProjectDetails"
      templateId:
        description: the project template identifier
        type: number
        format: long

      createdAt:
        type: string
        description: Datetime (GMT) when task was created
        readOnly: true
      createdBy:
        type: integer
        format: int64
        description: READ-ONLY. User who created this task
        readOnly: true
      updatedAt:
        type: string
        description: READ-ONLY. Datetime (GMT) when task was updated
        readOnly: true
      updatedBy:
        type: integer
        format: int64
        description: READ-ONLY. User that last updated this task
        readOnly: true
...
  ProjectResponse:
    title: Single project object
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          content:
            $ref: "#/definitions/Project"
...
  ProjectListResponse:
    title: List response
    type: object
    properties:
      id:
        type: string
        readOnly: true
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            type: array
            items:
              $ref: "#/definitions/Project"
  ProjectTemplateRequest:
    title: Project template request object
    type: object
    required:
      - name
      - key
      - category
      - scope
      - phases
    properties:
      name:
        type: string
        description: the project template name
      key:
        type: string
        description: the project template key
      category:
        type: string
        description: the project template category
      scope:
        type: object
        description: the project template scope
      phases:
        type: object
        description: the project template phases
  ProjectTemplateBodyParam:
    title: Project template body param
    type: object
    required:
      - param
    properties:
      param:
        $ref: "#/definitions/ProjectTemplateRequest"
...
  ProjectTemplateResponse:
    title: Single project template response object
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            $ref: "#/definitions/ProjectTemplate"
  ProjectTemplateListResponse:
    title: Project template list response object
    type: object
    properties:
      id:
        type: string
        readOnly: true
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            type: array
            items:
              $ref: "#/definitions/ProjectTemplate"
...
  ProductTemplateBodyParam:
    title: Product template body param
    type: object
    required:
      - param
    properties:
      param:
        $ref: "#/definitions/ProductTemplateRequest"
...
  ProductTemplateResponse:
    title: Single product template response object
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            $ref: "#/definitions/ProductTemplate"
  ProductTemplateListResponse:
    title: Product template list response object
    type: object
    properties:
      id:
        type: string
        readOnly: true
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            type: array
            items:
              $ref: "#/definitions/ProductTemplate"
...
  ProjectTypeRequest:
    title: Project type request object
    type: object
    required:
      - displayName
    properties:
      displayName:
        type: string
        description: the project type display name
  ProjectTypeBodyParam:
    title: Project type body param
    type: object
    required:
      - param
    properties:
      param:
        $ref: "#/definitions/ProjectTypeRequest"
...
  ProjectTypeCreateBodyParam:
    title: Project type creation body param
    type: object
    required:
      - param
    properties:
      param:
        $ref: "#/definitions/ProjectTypeCreateRequest"
...
  ProjectTypeResponse:
    title: Single project type response object
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            $ref: "#/definitions/ProjectType"
  ProjectTypeListResponse:
    title: Project type list response object
    type: object
    properties:
      id:
        type: string
        readOnly: true
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            type: array
            items:
              $ref: "#/definitions/ProjectType"
...
  ProductCategoryBodyParam:
    title: Product category body param
    type: object
    required:
      - param
    properties:
      param:
        $ref: "#/definitions/ProductCategoryRequest"
...
  ProductCategoryCreateBodyParam:
    title: Product category creation body param
    type: object
    required:
      - param
    properties:
      param:
        $ref: "#/definitions/ProductCategoryCreateRequest"
...
  ProductCategoryResponse:
    title: Single product category response object
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            $ref: "#/definitions/ProductCategory"
  ProductCategoryListResponse:
    title: Product category list response object
    type: object
    properties:
      id:
        type: string
        readOnly: true
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            type: array
            items:
              $ref: "#/definitions/ProductCategory"
...
  TimelineBodyParam:
    title: Timeline body param
    type: object
    required:
      - param
    properties:
      param:
        $ref: "#/definitions/TimelineRequest"
...
  TimelineResponse:
    title: Single timeline response object
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            $ref: "#/definitions/Timeline"
  TimelineListResponse:
    title: Timeline list response object
    type: object
    properties:
      id:
        type: string
        readOnly: true
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            type: array
            items:
              $ref: "#/definitions/Timeline"
...